Added additional code to function
Here i called the setNotes function  to update the state, and passed in a new array first item being the note you just created. Then the rest of the note will be passed in using the spread operator. then  we reset both the title and content states to be empty strings for good user experience