Created new function to handle noted the user clicked
Stored the clicked note into state. The type is going to of type Note. Then i created a function that handles the event when a user clicks on a note. This function is going to accept a note that the user clicks. the 'setSelectedNote' will add the note the user clicks. Then the title and content will also set to whatever content the selected note currently clicked