Adding new function handleUpdateNote
This prevents the form from submitting whenever the user clicks the save button.

Also created a new note object called updatedNote. Here the id will be set to the id of the selected note. Then the title and content will be selected from our form.

Then these values will be used in the updatedNotesList using a map function. This will then create a new array that will be updated with the new note object. As the map function iterates over the array of notes, we will check to see if the id of the note.id is = to the selectedNote.id that the user selects. If it is then we return the updated note otherwise we return the current note.

Once we have the update array we the n set all the objects to state. Which resets the form.