Added handleCancel function and additional buttons
Also added buttons for Save and Cancel.  Used conditional logic to render the different buttons based on if we have a selected note or not.

Also updated onSubmit in form to use a conditional statement. If we have a selected note then we call the handleUpdateNote function otherwise we add the handleAddNote function