Added deleteNote function
This function will take the event and the noteId to know which ID the user has decided to delete.

Then we will use Event.stopPropagation. Because our delete button is within Note and we have added an onClick event to the note itself, it stops any other event from interfering with the original onClick event, This is only necesarry if you have nested onClick events.

We then use filter function in updatedNotes to return only the ID of the note selected. Then we save the state.